+++ final/histogram-openmp.c
@@ -12,7 +12,7 @@ int main(void){
 	int frequencies[127];
 	#pragma omp for reduce(frequencies) local(i)
 	for(i = 0; i < N; i++)
-		frequencies[buffer[i]]++;
+		frequencies[(int) buffer[i]]++;
 
 	for(i = 32; i < 127; i++)
 		printf("%c: %d\n", i, frequencies[i]);
